Move pathanalysis to seperate files
